src/flora_identification/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
timm